requirements.txt
jsonpickle
trio==0.14.0
typing_extensions; python_version < '3.9'
aiohttp==3.9rc0
pyipv8
pyyaml